Initial setup for windows file attributes fix
